feat: extract link generation logic into a separate utility function
